// src/common/types.ts
export type UPermission = ValidPermissionKey<PermissionKey>
// ...
export type SettingBody = {
	value: string
	isSecret: boolean
	type: SETTING_DATA_TYPE
	description?: string
}
export type PermissionObj = Record<
	string,
	Record<
		string,
		{
			description?: string
			roles: string[]
		}
	>
>

export interface ModuleOptions<
	T extends Record<string, string> = typeof SETTING,
	TRU extends ReqUser = ReqUser,
	TUR extends UserResult = UserResult,
> {
	appSetting?: T
	appDefaultSettings?: Partial<Record<keyof T, SettingBody>>
	appPermissions?: PermissionObj
	getReqUser?: (currentUser: ReqUser) => Promise<TRU>
	getResultUser?: (currentUser: UserResult) => Promise<TUR>
}

export interface IErrorRes {
	code: string
	t: Date
	message: string
	statusCode: HttpStatus
	errors?: unknown
}

export interface ICursorPagingData<T> {
	docs: T[]
	hasNext: boolean
	nextCursor?: string
}

export interface IPagingData<T> {
	docs: T[]
	count: number
}

export interface ReqUser
	extends Pick<
		User,
		| 'id'
		| 'username'
		| 'mfaTelegramEnabled'
		| 'mfaTotpEnabled'
		| 'totpSecret'
		| 'telegramUsername'
		| 'enabled'
		| 'created'
		| 'modified'
	> {
	sessionId: string
	password: string
	permissions: UPermission[]
}

export type UserResult = Omit<
	ReqUser,
	'sessionId' | 'password' | 'totpSecret'
> & {
	telegramUsername: string | null
	permissions: UPermission[]
}